Just give up on isolating db
