+++ src/data.go
@@ -1,5 +1,9 @@
 package main
 
+import (
+	"gopkg.in/mgo.v2/bson"
+)
+
 type SiteInfo struct {
 	Title              string
 	Secret             string
@@ -22,22 +26,56 @@ type AdminConfig struct {
 	Admins       map[string]string
 }
 
-type Database interface {
-	Init(string, string) Database
-	Close()
-	NewThread(User, string, string, []string) (string, error)
-	ReplyThread(*Thread, User, string) (int, error)
-	GetThreadList(string, int, int, []string) ([]Thread, error)
-	GetThread(string) (Thread, error)
-	GetThreadById(interface{}) (Thread, error)
-	GetPost(interface{}) (Post, error) // Consider deprecating
-	GetPosts(*Thread, int, int) ([]Post, error)
-	PostCount(*Thread) (int, error)
-	GetPopularTags(int, int, []string) ([]Tag, error)
-	IncTag(string, interface{}) error // Move to internal only..?
-	DecTag(string) error              // Move to internal only.. ?
-	EditPost(interface{}, string) error
-	DeletePost(interface{}, bool) error
-	SetThreadTags(interface{}, []string) error
-	GetMatchingTags(string, int, int, []string) ([]Tag, error)
+type User struct {
+	Nickname string
+	Tripcode string
+}
+
+type Thread struct {
+	Id         bson.ObjectId "_id"
+	ShortUrl   string
+	Title      string
+	Author     User
+	Tags       []string
+	Date       int64
+	Messages   int32
+	ThreadPost bson.ObjectId
+	LastReply  bson.ObjectId
+	LRDate     int64
+}
+
+type Post struct {
+	Id           bson.ObjectId "_id"
+	ThreadId     bson.ObjectId
+	Author       User
+	Content      string
+	Date         int64
+	LastModified int64
+	ContentType  string
+}
+
+type Counter struct {
+	Name string
+	Seq  int64
+}
+
+type Tag struct {
+	Name       string
+	Posts      int64
+	LastUpdate int64
+	LastThread bson.ObjectId
+}
+
+type TagData struct {
+	Name       string
+	LastUpdate int64
+	LastThread ThreadInfo
+	LastIndex  int64
+}
+
+type ThreadInfo struct {
+	Thread      Thread
+	LastPost    Post
+	LastMessage int
+	Page        int
 }
